Rename hook strategies to hook actions and make their modules package private
