+++ guet/commands/usercommands/start/_cancel_start_strategy.py
@@ -1,7 +1,7 @@
-from guet.commands.usercommands.start.hook_strategy import HookStrategy
+from guet.commands.usercommands.start._hook_action import HookAction
 
 
-class CancelStartStrategy(HookStrategy):
+class CancelStartAction(HookAction):
     def _hook_apply(self) -> None:
         pass
