+++ guet/commands/usercommands/start/_create_alongside_hook_strategy.py
@@ -1,0 +1,6 @@
+from guet.commands.usercommands.start._hook_action import HookAction
+
+
+class CreateAlongsideHookAction(HookAction):
+    def _hook_apply(self) -> None:
+        self.git.create_hooks(alongside=True)
