+++ test/commands/usercommands/start/test_create_alongside_hook_strategy.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 from unittest.mock import patch, Mock
 
-from guet.commands.usercommands.start.create_alongside_hook_strategy import CreateAlongsideHookStrategy
+from guet.commands.usercommands.start._create_alongside_hook_strategy import CreateAlongsideHookAction
 
 
 class TestCreateAlongsideHookStrategy(TestCase):
@@ -9,6 +9,6 @@ class TestCreateAlongsideHookStrategy(TestCase):
     def test_creates_hooks_with_alongside_mode(self):
         context = Mock()
         context.git = Mock()
-        strategy = CreateAlongsideHookStrategy(context.git)
+        strategy = CreateAlongsideHookAction(context.git)
         strategy.apply()
         context.git.create_hooks.assert_called_with(alongside=True)
